optimized by Claude for faster processing + lower memory usage

+++ PIDTesting-2yoke4coil/Controller.hpp
@@ -100,6 +100,6 @@ class FullController {
     int16_t FRPrev;
     int16_t BRPrev;
 
-    unsigned int tDiff;
+    float tDiff;
 };
 #endif // CONTROLLER_HPP
